requirements.txt
# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# AI/ML
openai==1.12.0
pydantic==2.6.1

# Data processing
python-dotenv==1.0.1

# Utilities
tqdm==4.66.1  # Progress bars (optional, для улучшения UX)

